Damage Range and Character Creation Updates
Updated the damage the player deals so that it is now a range based on their level and ap stats rather than just a set number.

Made some adjustments on how the AP is stored and modified. Makes it a lot easier to deal with than the earlier system.

Character Creation ability points now start at 0. Also added an info box that explains the difference between rolling and incrementing/decrementing the AP.

Changed the text size for the player name so that longer names fit.

Made it so that items don't collide with each other.