Weapon Equips and Damage Range UI fix
Weapons can now be equipped. Your damage range now increases based on the damage of the weapon. It also shows in the appropriate slot of the Character screen.

Fixed a couple of bugs regarding the damage range in the Character screen UI. If you cancelled the AP changes, your stats would return to the original stats from level 1 instead of where it was before you made the changes. Another bug was that the damage range would not return to the original range when cancelling AP changes. These have both been fixed.

Also did a little bit of folder management so the hierarchy should be less confusing.