Teleporting and Enemy Death Animation
Added a "platform" that allows the player to teleport to the next map. Very basic at this point, but is working well. Each teleporting platform has an ID and finds its partner with a matching ID when the map loads. When the player touches the platform and presses the Up key, they will move to its partner platform and the camera will snap to the location as well. Might cause some loading slowdowns in the future once more areas are added (going through each individual teleport platform might end up being costly), but should be straightforward to adjust in the future.

Death animations will now play when you kill an enemy. This also fixes a bug where your damage would not show on the last hit (the enemy gameobject was destroyed immediately, so the damage would not pop up).