Hotkey Changes & Starting on Dialogue/Skills
Made some changes to how hotkeys are organized and interact with objects. The amount of an item you have will now update in the hotkey slots. So picking up and using items will be consistent in both the inventory and hotkey slots.

Started working on the UI for Dialogue and Skills. Nothing concrete yet, but have some very basic screens up and running to place some prefabs when they are created.