Single target skills!
A system for single target skills has been properly set up. It will attack a target within a set range for a specifc percentage of your damage. It will show the skill hitting the enemy, damage them appropriately, and updates the exp gained if they are killed. I should also be able to just attach an animation to the skill prefab so that will play instead of just an icon over the player.

Still need to do for this: Skill UI, showing the exp earned on the Gains UI, and dragging and dropping skills to the hotkey slots.

For now, since I am still testing and playing around with it, you just press 'Y' and it will attack the enemy with a Double Slash skill.