Hotkey dragging quick fix
Found the main reason why dragging and dropping items/skills into the hotkey slots was so difficult. It was not stopping when it hit the first overlap, so it would continue to go through the list and choose the very last one. Still need to do some tweaking, but this has solved a good deal of the issues and feels a bit more intuitive.