Camera Follow, Sprite Adjustments, Teleport Transition Effects, Item Drop Adjustment
The camera will now follow the player as you move throughout the map. The speed and offset can be adjusted so I will probably be tweaking this a bit in the near future.

Some minor adjustments to the sprites have been made to prevent weird blurriness and scaling issues. Also fixed some of the jittering when moving the player.

The screen will now fade out to black when teleporting and fades back in when the player reaches the new area.

A small adjustment to item drops. I applied a minor force to the items so it will look more like they are exploding out of the enemies.