New Sprites and New Physics System
Created a new spriteset for the platforms and a background so that it's not so bare. This allows me to create different types of platforms and customize the map. I also implemented a quick tool that makes the platforms snap together to make life a whole lot easier.

I have also been working on implementing my own physics system, specifically gravity, adding forces, and collision detection. Using Unity's built in physics and collision was causing a lot of problems and just felt like I was trying to fight it to do what I want. So far, using my new system I can move the player, collide with platforms (but still jump through them), and climb ropes. This system still uses Unity 2D colliders to make it easier to set up and use bounds, but avoids rigidbodies and the rest of the physics system.

Next steps are:
1. Tweak bottom collision a bit so that the player more accurately and consistently stands on top of platforms.
2. Add collision detection on the sides
3. Test physics system on enemies
4. Refactor the existing code that uses Unity physics (knockbacks, player & enemy damage, item drops) to use my new system