Physics Updates
Left and right collision added. New physics system applied to Enemies and items. Will take damage from touching enemies and can pickup items when colliding with them.

Refactored some of the collision detection so that it only directly affects the object if the bottom is touching a platform. The rest of the collisions (with enemies, ropes, items, etc) are stored and can be used by the appropriate controllers.