Skill Descriptions and MP usage
Clicking on a skill now shows its information in the Skills UI. This includes the sprite, name of the skill, required level, amount of mana it uses, and a short description of what it does.

Using skills now uses mana. You cannot use a skill if you don't have enough mana.