Working on Dialogue
Started doing more work on the Dialogue system. So far, I have it set up in three separate parts: Dialogue Lines, Dialogue Scenes, and Dialogue Responses.

A Dialogue line contains a single monologue, for lack of a better word, that an NPC would say. It also contains a linked Dialogue Response for the player.
A Dialogue Response is essentially what the player would click to respond. For now it is set to 3 different types of responses: End, Continue, or Accept. These responses, in the future, will perform an action when clicked. End will end the chat, Continue will go to the next line of dialogue, and Accept will accept a quest.
A Dialogue Scene is essentially a collection of Dialogue Lines. It stores a text file and splits it into different lines based on a delimiter. The file also has certain flags at the end of each line to set up the type of response you want available to the player.