More on Dialogue System
Today, I've been working on making updates to the dialogue system.

Each NPC now has a unique list of Dialogue Scenes. All you have to do to add a new Dialogue Scene to an NPC is increase the size of the list and drag a text file to it through the editor. The already existing dialogue code will take care of everything else (loading the file, splitting between different scenes, flags, etc). A new flag has also been added ([TITLE]) for the title of the Dialogue scene. This is essentially like the name of a quest so that the player can pick and choose which dialogue they want to start.

 A new UI element has been added that will display the different Dialogue Scenes when double clicking on an NPC. They are currently set up as buttons so that the player can just easily click on the one they want.

Next steps:
1. Starting the dialogue when clicking on an option in the new UI.
2. Mapping the flags to buttons that will continue and end dialogue.
3. Create and add portrait sprites to display when talking to an NPC.