Dialogue Reworks/Updates
I reworked the way that the NPC Dialogues are loaded in. Instead of attaching each relevant dialogue scene to its appropriate NPC, I just created a big list in the Game Controller. I added a new flag in the text reader that will allow it to properly assign the dialogue scene to its NPC and each scene is given its own ID.
Now, clicking on a dialogue option will close the selection screen, open the actual dialogue screen, and display the line of dialogue. The response button will update its text according to the flag set in the dialogue line and will either close or continue the dialogue.

I also made a quick update to the physics object script that will snap the object to the top of a platform when it lands. This has fixed any inaccuracies when jumping and falling (objects used to land slightly above or inside the platforms depending on the speed at which they fell).